+++ packages/eslint-config-airbnb-base/legacy.js
@@ -33,6 +33,7 @@ module.exports = {
       },
     ],
     'no-var': 'off',
+    'prefer-object-spread': 'off',
     strict: ['error', 'safe'],
   },
 };